# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/departments": {
            "get": {
                "tags": [
                    "departments"
                ],
                "description": "Returns all departments",
                "operationId": "findDepartments",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorts items base of field provided",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "-id",
                                "created_at",
                                "-created_at",
                                "department",
                                "-department"
                            ]
                        },
                        "examples": {
                            "-id": {
                                "summary": "sorts in an descending manner by append a '-' ",
                                "value": "-id"
                            },
                            "id": {
                                "summary": "sorts in an ascending manner",
                                "value": "id"
                            }
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filters base on the parameter passed e.g filter[faculty_id]=5",
                        "required": false,
                        "schema": {
                            "properties": {
                                "faculty_id": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "$ref": "#components/parameters/range"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success respones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Department"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requested filter(s) `description, are not allowed. Allowed filter(s) are `created_at` "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/departments/{id}": {
            "get": {
                "tags": [
                    "departments"
                ],
                "description": "Returns a department based on a single ID",
                "operationId": "findDepartmentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of department to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "single department response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Department"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/departments/{id}/courses": {
            "get": {
                "tags": [
                    "departments"
                ],
                "description": "Returns all course registered for a department",
                "operationId": "findDepartmentsCourses",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of department for which courses are to be fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "courses respones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Department"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Model] 2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/faculties": {
            "get": {
                "tags": [
                    "faculties"
                ],
                "description": "Returns faculties",
                "operationId": "findFaculties",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorts items base of field provided",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "-id",
                                "created_at",
                                "-created_at",
                                "faculty",
                                "-faculty"
                            ]
                        },
                        "examples": {
                            "-id": {
                                "summary": "sorts in an descending manner by append a '-' ",
                                "value": "-id"
                            },
                            "id": {
                                "summary": "sorts in an ascending manner",
                                "value": "id"
                            }
                        }
                    },
                    {
                        "$ref": "#components/parameters/range"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success respones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Faculty"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requested filter(s) `description, are not allowed. Allowed filter(s) are `created_at` "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/faculties/{id}": {
            "get": {
                "tags": [
                    "faculties"
                ],
                "description": "Returns a faculty based on a single ID",
                "operationId": "findFacultyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of faculty to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "single department response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Faculty"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Department": {
                "required": [
                    "department",
                    "faculty_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "department": {
                        "type": "string"
                    },
                    "banner": {
                        "type": "string"
                    },
                    "faculty_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Department"
                }
            },
            "Faculty": {
                "required": [
                    "faculty"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "faculty": {
                        "type": "string"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Faculty"
                }
            },
...
        },
...
    },
    "security": [
        []
    ]
}